Add authorization checks for Organisation APIs

+++ src/main/java/org/rutebanken/tiamat/service/organisation/OrganisationDeleter.java
@@ -1,6 +1,7 @@
 package org.rutebanken.tiamat.service.organisation;
 
 import com.google.api.client.util.Preconditions;
+import org.rutebanken.helper.organisation.ReflectionAuthorizationService;
 import org.rutebanken.tiamat.auth.UsernameFetcher;
 import org.rutebanken.tiamat.changelog.EntityChangedListener;
 import org.rutebanken.tiamat.lock.MutateLock;
@@ -18,6 +19,10 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 
+
+import static org.rutebanken.helper.organisation.AuthorizationConstants.ROLE_EDIT_STOPS;
+import static org.rutebanken.helper.organisation.AuthorizationConstants.ROLE_ORGANISATION_EDIT;
+
 @Service
 public class OrganisationDeleter {
 
@@ -31,15 +36,19 @@ public class OrganisationDeleter {
 
     private final MutateLock mutateLock;
 
+    private final ReflectionAuthorizationService authorizationService;
+
     @Autowired
     public OrganisationDeleter(OrganisationRepository organisationRepository,
                                EntityChangedListener entityChangedListener,
                                UsernameFetcher usernameFetcher,
-                               MutateLock mutateLock) {
+                               MutateLock mutateLock,
+                               ReflectionAuthorizationService authorizationService) {
         this.organisationRepository = organisationRepository;
         this.entityChangedListener = entityChangedListener;
         this.usernameFetcher = usernameFetcher;
         this.mutateLock = mutateLock;
+        this.authorizationService = authorizationService;
     }
 
     @Transactional
@@ -51,7 +60,7 @@ public class OrganisationDeleter {
 
             List<Organisation> organisations = getAllVersionsOfOrganisation(organisationId);
 
-            // TODO: would we need any authorization by role here?
+            authorizationService.assertAuthorized(ROLE_ORGANISATION_EDIT, organisations);
             organisationRepository.deleteAll(organisations);
             notifyDeleted(organisations);
